Leere Kommentare verhindern

+++ rezeptbuch/WebContent/js/viewRecipe.js
@@ -17,43 +17,47 @@ function init() {
 	
 }
 
-//Die sichtbare Ausführung ist wegen dem Versand von Mails leicht verzögert.
+// Die sichtbare Ausführung ist wegen dem Versand von Mails leicht verzögert.
 function addComment() {
 	var id = document.getElementById('id').value;
 	var comment = document.getElementById('newComment').value;
 
-	var xmlhttp = new XMLHttpRequest();
-	xmlhttp.onreadystatechange = function() {
-		if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
-			var commentResponse = JSON.parse(xmlhttp.responseText);
+	if (comment != "") {
+		var xmlhttp = new XMLHttpRequest();
+		xmlhttp.onreadystatechange = function() {
+			if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
+				var commentResponse = JSON.parse(xmlhttp.responseText);
 
-			var commentsTable = document.getElementById('commentsTable');
-			var nextRowPosition = commentsTable.rows.length;
+				var commentsTable = document.getElementById('commentsTable');
+				var nextRowPosition = commentsTable.rows.length;
 
-			var newTableRow = commentsTable.insertRow(nextRowPosition);
-			var newCommentAuthor = newTableRow.insertCell(0);
-			var newCommentComment = newTableRow.insertCell(1);
-			var newCommentDelete = newTableRow.insertCell(2);
+				var newTableRow = commentsTable.insertRow(nextRowPosition);
+				var newCommentAuthor = newTableRow.insertCell(0);
+				var newCommentComment = newTableRow.insertCell(1);
+				var newCommentDelete = newTableRow.insertCell(2);
 
-			newCommentComment.innerHTML = commentResponse.comment;
-			newCommentAuthor.innerHTML = commentResponse.author;
-			newCommentDelete.innerHTML = '<a href=\'/rezeptbuch/DeleteCommentServlet?id='
-					+ commentResponse.id
-					+ '&recipe='
-					+ commentResponse.recipe
-					+ '\'>Löschen</a>';
-			
-			document.getElementById('commentForm').reset();
+				newCommentComment.innerHTML = commentResponse.comment;
+				newCommentAuthor.innerHTML = commentResponse.author;
+				newCommentDelete.innerHTML = '<a href=\'/rezeptbuch/DeleteCommentServlet?id='
+						+ commentResponse.id
+						+ '&recipe='
+						+ commentResponse.recipe + '\'>Löschen</a>';
 
+				document.getElementById('commentForm').reset();
+
+			}
 		}
-	}
 
-	xmlhttp.open('POST', '../AddCommentServlet', true);
-	xmlhttp.setRequestHeader('Content-Type',
-			'application/x-www-form-urlencoded');
-	xmlhttp.send('recipe=' + id + '&comment=' + comment);
-	
-	
+		xmlhttp.open('POST', '../AddCommentServlet', true);
+		xmlhttp.setRequestHeader('Content-Type',
+				'application/x-www-form-urlencoded');
+		xmlhttp.send('recipe=' + id + '&comment=' + comment);
+
+	}
+	else{
+		window.alert("Bitte geben Sie ihren Kommentar in das Feld ein.");
+		
+	}
 }
 
 function aboHandling() {
